requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
python-dotenv==1.2.1
openai==1.30.1
httpx==0.27.2
